Stopgap help for the CPU-bound long hang in bug fdo#31716
As the hang seems to be causes by the code messing around in extension
help file directories, of which there are lots in a multi-lingual
build, just build en-US help for the Presenter Console
(a.k.a. presenter-screen) extension for now.

Why such a change is needed both here and in
solenv/inc/extension_helplink.mk (see commit
4e40f4ea586f240f38aa6856eda64619f505294a in the "bootstrap" repo) I
don't know, but that's how it seemed in my testing.